Clarified some sampling issues.
- Previously, the --gtyp-ppn-xxx-pre options were used in a way that let us believe that we were sampling individuals from the previous generation, but I have changed that now that we have a more explicit definition of the spip annual cycle as given in the about-spip vignette for CKMRpop.
- The --gtyp-ppn-xxx-post option now allows all ages
from 1 to MA to be sampled.  This fixes a bug that
formerly made it impossible to sample MA-year-olds
that could have reproduced that year.  Thanks to
Robin Waples for catching that bug.
-A new version is compiled for MacOSX now, too.

Note that the year subscripts for samples now differ
from what was reported previously so that the
examples, etc. included in this repo may no longer
match what they used to.